refactor(ui): extract ViewportTabBar from Viewport (#855)
Move the tab strip (term/todo/activity/diff + preview tab) and the whole
preview-start feature (popover state, arming, outside-click dismiss,
window listener) + scoped CSS into a child; tab is $bindable. The 3
cross-boundary .vp-head.mobile rules rewritten to .tab-group.mobile.
Viewport <template> 107/165 -> 98/152; child 10/13.
